AI/requirements.txt
setuptools
fastapi
pandas
numpy
scikit-learn
python-dotenv
seaborn
matplotlib
ipykernel
pymongo
uvicorn
requests
geopy
-e .